tests():  * changed bevy_component tests to use testing/bevy_example as source for registry data  * removed testing/bevy_registry_export  * related tweaks & cleanups
