--- testing/bevy_registry_export/basic/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "bevy_bevy_registry_export_basic_testing"
-version = "0.3.0"
-edition = "2021"
-license = "MIT OR Apache-2.0"
-
-[dependencies]
-bevy="0.12"
-bevy_gltf_blueprints = { path = "../../../crates/bevy_gltf_blueprints" }
-bevy_registry_export = { path = "../../../crates/bevy_registry_export" }
-bevy_gltf_worlflow_examples_common = { path = "../../../examples/common" }
-
-bevy_rapier3d = { version = "0.23.0", features = [ "serde-serialize", "debug-render-3d", "enhanced-determinism"] }
-bevy_asset_loader = { version = "0.18", features = ["standard_dynamic_assets" ]} 
-bevy_editor_pls = { version = "0.6" }
-rand = "0.8.5"
