--- testing/bevy_registry_export/basic/assets/assets_core.assets.ron
@@ -1,1 +1,0 @@
-({})
