--- testing/bevy_registry_export/basic/assets/assets_game.assets.ron
@@ -1,6 +1,0 @@
-({
-    "world":File (path: "models/World.glb"),
-    "models": Folder (
-        path: "models/library",
-    ),
-})
