+++ tools/bevy_components/tests/test_registry.py
@@ -1,11 +1,10 @@
 import bpy
+from .setup_data import setup_data
 
-
-def test_blend():
+def test_blend(setup_data):
     registry = bpy.context.window_manager.components_registry
-    registry.schemaPath = "../../testing/bevy_registry_export/basic/assets/registry.json"
+    registry.schemaPath = setup_data["schema_path"]
     bpy.ops.object.reload_registry()
-    #print("registry type infos", registry.type_infos)
 
     short_name = "BasicTest"
     component_type = registry.short_names_to_long_names[short_name]
@@ -19,11 +18,6 @@ def test_blend():
     target_components_metadata = object.components_meta.components
     component_meta = next(filter(lambda component: component["name"] == short_name, target_components_metadata), None)
     propertyGroup = getattr(component_meta, property_group_name, None)
-    print("propertyGroup", propertyGroup, propertyGroup.field_names)
-
-
-    """copy_component_operator = bpy.ops.object.copy_bevy_component
-    copy_component_operator()"""
 
 
     assert propertyGroup.field_names == ['a', 'b', 'c']
